+++ packages/create-faas-app/src/action.ts
@@ -34,7 +34,7 @@ function buildPackageJSON(name: string): string {
       scripts: {
         dev: 'vite',
         build: 'vite build',
-        start: 'node server.ts',
+        start: 'node --import @faasjs/node-utils/register-hooks server.ts',
         check: 'faas lint',
         test: 'vitest run',
         'migrate:latest': 'faas knex latest',