+++ packages/node-utils/src/register_hooks.ts
@@ -1,0 +1,3 @@
+import { registerNodeModuleHooks } from './load_package'
+
+registerNodeModuleHooks()